requirements.txt
aws-cdk-lib==2.47.0
boto3
constructs>=10.0.0,<11.0.0
pytest==6.2.5
scikit-learn==0.23.1
